requirements/dev.txt
-r base.txt
django-debug-toolbar==4.2.0
# Required by django-debug-toolbar
sqlparse==0.4.4

factory-boy==3.3.0
coverage==7.3.0
flake8==6.1.0
# Required by flake8
pyflakes==3.1.0  # constrained by flake8
pep8==1.7.1
mccabe==0.7.0

# For translation
transifex-client==0.11b3

requires.io

# For docs
Sphinx==1.3.6
sphinx-rtd-theme==0.1.9
alabaster==0.7.7
Babel==2.2.0
docutils==0.12
Pygments==2.16.1
snowballstemmer==1.2.1

# For debugging
ipdb